backend/pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bescosial-context-tester"
version = "0.1.0"
description = "GUI for testing Vertex AI summaries on dialog JSON"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "python-multipart>=0.0.9",
  "orjson>=3.9",
  "google-auth>=2.27",
  "google-cloud-aiplatform>=1.41",
  "requests>=2.31"
]
